Cache parser in JSGLRParseService
The JSGLRI classes no longer take the source file and the input string
as parameters to the constructor, but as parameters to `parse` instead.
This makes sure that entire parsers can be cached, instead of just their
configurations.
Because of this change, the incremental parser can store intermediate
results in its local state.