Fix NPE
Whenever there are input units to multi file analysis that do not have a source, a message was logged that these would be ignored. However, the code did not actually ignore them, which caused a NullPointerException when trying to get a resource key for this unit (old line 132, new line 133).